+++ src/app/modals/projects-create-modal/projects-create-modal.component.ts
@@ -99,14 +99,16 @@ export class ProjectsCreateModalComponent implements OnInit {
     this.apiDataProjectService.createProject(project, this.userID)
     .subscribe(
       res => {
-        // console.log(res);
         project.projectID = res.newProjectID;
         this.createSuccess.emit(project);
         this.websocketService.sendNewProject(res);
         // clear the projects browse data in the cache, to force the projects search component to refresh from the database
         this.cacheService.projectsBrowseData = undefined;
+
+        this.cacheService.raiseToast('success', `Project ${project.projectName} created`);
       },
       err => {
+        this.cacheService.raiseToast('error', err);
         // console.log(err);
       }
     );